requirements.txt
fastapi
uvicorn
transformers
torch
streamlit
requests
google-generativeai